backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
firebase-admin
google-generativeai
google-cloud-documentai
pypdf
pydantic
